+++ src/riley/zig/scalingpolicy.zig
@@ -69,7 +69,7 @@ pub fn tileSize(
     tile_size_min: u16,
     tile_size_max: u16,
     pixels_num: [2]u32,
-    sub_sample: u16,
+    sub_sample: u32,
     halo_px: u16,
 ) u16 {
     if (tile_size_override) |tile_size| {
